converting to be phonegap 3.0 compatible
